Added NT Hash option
This allows the user to specify an NTDS dump instead of a cracked password file with the '-t' option.  This adds an nthash field to each object and their NT Hash value.  Useful for having instant access to the hash of a user for PtH, and will also allow for queries that can match accounts with shared passwords, even if you couldn't crack them.